- fix tests die on uninited mutex - vm_map init seq was wrong - added test (runs many sleeping threads) - on 1000 threads kernel is getting crazy. now limited to 200, TODO FIXME

+++ trunk/oldtree/kernel/phantom/unix/fs_dev.c
@@ -229,7 +229,7 @@ static uufile_t *  dev_namei(uufs_t *fs, const char *filename)
 #define NPART 10
 
         const char *oname[NPART];
-        int olen[NPART];
+        size_t olen[NPART];
 
         int nb = uu_break_path( filename, NPART, oname, olen );
